QuickPlan-main/requirement.txt
asgiref==3.6.0
boto3==1.26.100
botocore==1.29.100
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.18
django-crontab==0.7.1
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
numpy==1.24.2
pandas==2.0.0
Pillow==9.5.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
